Develop

+++ src/main/java/kr/co/lotteon/service/OrderService.java
@@ -230,7 +230,7 @@ public class OrderService {
         if ("type4".equals(orderPay)){
             orders.put("orderStatus", "입금대기");
         }else {
-            orders.put("orderStatus", "결제완료");
+            orders.put("orderStatus", "배송준비");
         }
 
         Orders ordered = saveOrders(orders);
@@ -249,7 +249,7 @@ public class OrderService {
             if ("type4".equals(orderPay)){
                 orderDetail.put("detailStatus", "입금대기");
             }else {
-                orderDetail.put("detailStatus", "상품준비");
+                orderDetail.put("detailStatus", "배송준비");
             }
 
             orderDetail.put("detailPoint", String.valueOf(detailPoint));
@@ -257,7 +257,9 @@ public class OrderService {
 
             int cartProdNo = Integer.parseInt(orderDetail.get("cartProdNo"));
             // cart 테이블에서 삭제
-            cartProductRepository.deleteById(cartProdNo);
+            if (cartProdNo != 0){
+                cartProductRepository.deleteById(cartProdNo);
+            }
 
             int prodNo = Integer.parseInt(orderDetail.get("prodNo"));
             int optNo = Integer.parseInt(orderDetail.get("optNo"));
